Use underscore prefixes only for special cases
Underscore-prefix private property names only if they are mutable private properties with a similarly named internal property.

This is more Swifty, and we're already separating private methods and properties through the file organization as well as the access control declarations.

The only place where the underscore prefix makes sense is in the special case where you have identically named private and public properties, with the public property exposing a static copy of the privately mutable property.