Avoid staircase of illegibility (#31)
* Avoid staircase of illegibility

Updated 2.4 to clarify how to avoid the "pyramid of illegibility" (or staircase) auto indentation problems when a function has external and internal parameters.

Xcode 7.3 no longer indents incorrectly when the opening brace is on the last line of the declaration, so I removed those comments. I replaced them with a comment about how the next line could blend in with the previous, which is something we probably want to avoid anyways. So although Xcode fixes this issue, we probably want to keep the convention for this purpose.

* Updated 2.4 with examples when first parameter has an external name, and updated 2.5 for Xcode 7.3 indentation logic.