Add another style guide goal
One of the major benefits in having a style guide (especially if we have "fix it" actions), is that when reviewing someone's pull request, you don't have to make any comments about the style. It'll presumable already be of the agreed upon style and you can just focus on the actual logic of what they're implementing.