Add unique identifiers to each rule
This removes the (link) from every rule and replaces it with a unique identifier (following the numerical conventions from our Javascript style guide https://github.com/airbnb/javascript)

The benefit of doing this is that when you reference a style guide rule, you can reference it by number instead of pasting in the entire rule / a sentence. e.g. "This violates rule 3.4 of the style guide" instead of "This violates 'long function invocations should also break on each argument'"