Clarify Rule: Never use the default case when switching over an enum
Summary
This PR proposes clarifying the rule "Never use the default case when switching over an enum" to be more specific.

Reasoning
I clarified rule with "real world" example and to follow clear and understandable examples that the guide is providing. (such as universe, weather and etc.)

Open to suggestions on this, thanks! 😄

Reviewers
cc. @calda

Please react with 👍/👎 if you agree or disagree with this proposal.