Build MVP website from README (#338)
#### Summary

Adds the ability to convert the README.md to a website. This
productionizes https://github.com/airbnb/swift/pull/336 following
positive internal feedback on the direction.

The next steps will be to set up a GitHub workflow to publish the site
to GitHub Pages. We may also want to continue to polish the style.

#### Reasoning

A website will make the style guide more accessible. We chose to
generate the website from the README so that the existing workflow for
updating the style guide remains the same.


#### Testing

- [x] Run `bundle exec rake site:serve` locally and check the content at
http://localhost:4000
- [x] Verify that top badges and SPM command plugin sections are
filtered out of the site

| Top | Code block | Footer |
| --- | --- | --- |

|![image](https://github.com/user-attachments/assets/176cd61e-fb77-40ef-aa5c-8d25858eb883)
|
![image](https://github.com/user-attachments/assets/1c080b2a-2edd-4d7a-a674-ab1983cb61d4)
|
![image](https://github.com/user-attachments/assets/0f8bc71e-04ae-4719-97e3-7595635f6a18)
|


#### Please review

@airbnb/swift-styleguide-maintainers